move uploader-related code to uploader.py
